Add a fallback to find show_id
First, instead of trying to find exactly the show in the search, settle
on a name that starts with the show title we have found.
If it can't be found, look at the list of all the shows and try to find
the show name (exactly this time).

If it's still is not found, well... no luck !