myepisodes: Sanitize show name found
Sanitize the link we find on the website if the title we are using comes
from a filename.

The rationale behind this is that filename contains weird character
sometimes, but the link on MyEpisodes.com still may contains them, so we
have to sanitize both in order to compare them on the same basis.